Removed no longer used parts, optimized and reogrganized messy code.
