+++ conf/CRD3Dataset_val.yaml
@@ -1,20 +1,8 @@
-# This file houses the configuration for the CRD3 data.
-
-# Path to the directory containing aligned summary-dialogue files
-# This value assumes you are running from the src/ directory. Adjust as needed.
-CRD3_path: '../CRD3/data/aligned data/c=4/'
-# Number of aligned data files into which each episode is split. This corresponds to c=n in the CRD3 path
-files_per_episode: 8
-# Path to the file containing the stems of train files
-idx_file: '../data/val_files'
-# Number of summaries present in the datasets (to avoid calculation of dataset length)
-n_summary: ''
-# Size of the chunk buffer of the dataset. Large buffers will allow for randomization over more files, but will consume
-# more memory.
-buffer_size: 32
-# Path to text vocab
-tokenizer_path: '../data/tokenizer.json'
-# Maximum length of source sequences
-max_seq_len: 4096
-# Maximum length of target (summary) sequences
-max_tgt_seq_len: 200
+max_seq_len: 2048
+max_tgt_seq_len: 512
+prompt_prefix: "The following is an excerpt from a show in which multiple actors are playing Dungeons and
+Dragons:\n\n"
+prompt_suffix: "\nThat is the end of the script.\n\
+Write a summary of the script, making sure to include all the major events contained within.\n"
+split: "validation"
+tokenizer_path: "unsloth/mistral-7b-bnb-4bit"
