--- data/check_files
@@ -1,1 +1,0 @@
-C1E001
