--- data/speaker_tokenizer.json
@@ -1,124 +1,0 @@
-{
-  "version": "1.0",
-  "truncation": null,
-  "padding": null,
-  "added_tokens": [
-    {
-      "id": 0,
-      "content": "<UNK>",
-      "single_word": false,
-      "lstrip": false,
-      "rstrip": false,
-      "normalized": false,
-      "special": true
-    }
-  ],
-  "normalizer": {
-    "type": "Sequence",
-    "normalizers": [
-      {
-        "type": "NFKC"
-      },
-      {
-        "type": "BertNormalizer",
-        "clean_text": true,
-        "handle_chinese_chars": true,
-        "strip_accents": null,
-        "lowercase": true
-      }
-    ]
-  },
-  "pre_tokenizer": {
-    "type": "BertPreTokenizer"
-  },
-  "post_processor": null,
-  "decoder": {
-    "type": "CTC",
-    "pad_token": "<pad>",
-    "word_delimiter_token": "|",
-    "cleanup": true
-  },
-  "model": {
-    "type": "WordLevel",
-    "vocab": {
-      "<UNK>": 0,
-      "matt": 1,
-      "laura": 2,
-      "sam": 3,
-      "marisha": 4,
-      "travis": 5,
-      "liam": 6,
-      "taliesin": 7,
-      "ashley": 8,
-      "orion": 9,
-      "mary": 10,
-      "will": 11,
-      "zac": 12,
-      "felicia": 13,
-      "deborah": 14,
-      "all": 15,
-      "jon": 16,
-      "khary": 17,
-      "joe": 18,
-      "sumalee": 19,
-      "wil": 20,
-      "mark": 21,
-      "noelle": 22,
-      "darin": 23,
-      "chris": 24,
-      "hardwick": 25,
-      "patrick": 26,
-      "rothfuss": 27,
-      "kit": 28,
-      "brian": 29,
-      "audience": 30,
-      "ivan": 31,
-      "hector": 32,
-      "robbie": 33,
-      "becca": 34,
-      "-": 35,
-      "ify": 36,
-      "lucas": 37,
-      "crew": 38,
-      "member": 39,
-      "joel": 40,
-      "perkins": 41,
-      "amy": 42,
-      "lenore": 43,
-      "sean": 44,
-      "sound": 45,
-      "yasmine": 46,
-      "yuri": 47,
-      "ryan": 48,
-      "offscreen": 49,
-      "bag": 50,
-      "in": 51,
-      "toy": 52,
-      "ahsley": 53,
-      "amandine": 54,
-      "babs": 55,
-      "kai": 56,
-      "max": 57,
-      "and": 58,
-      "eric": 59,
-      "alec": 60,
-      "cast": 61,
-      "denise": 62,
-      "erika": 63,
-      "jore": 64,
-      "producer": 65,
-      "allison": 66,
-      "annet": 67,
-      "cameraman": 68,
-      "computer": 69,
-      "generated": 70,
-      "iphone": 71,
-      "joey": 72,
-      "screen": 73,
-      "steve": 74,
-      "voice": 75,
-      "willmott": 76
-    },
-    "unk_token": "<unk>"
-  }
-}
