--- data/test_files
@@ -1,16 +1,0 @@
-C1E108
-C1E109
-C1E110
-C1E111
-C1E112
-C1E113
-C1E114
-C1E115
-C2E039
-C2E040
-C2E041
-C2E042
-C2E043
-C2E044
-C2E045
-C2E046
