--- data/train_files
@@ -1,139 +1,0 @@
-C1E001
-C1E002
-C1E003
-C1E004
-C1E005
-C1E006
-C1E007
-C1E008
-C1E009
-C1E010
-C1E011
-C1E013
-C1E014
-C1E015
-C1E016
-C1E017
-C1E018
-C1E019
-C1E020
-C1E021
-C1E022
-C1E023
-C1E024
-C1E025
-C1E026
-C1E027
-C1E028
-C1E029
-C1E030
-C1E031
-C1E032
-C1E033
-C1E034
-C1E035
-C1E036
-C1E037
-C1E038
-C1E039
-C1E040
-C1E041
-C1E042
-C1E043
-C1E044
-C1E045
-C1E046
-C1E047
-C1E048
-C1E049
-C1E050
-C1E052
-C1E053
-C1E054
-C1E055
-C1E056
-C1E057
-C1E058
-C1E059
-C1E060
-C1E061
-C1E062
-C1E063
-C1E064
-C1E065
-C1E066
-C1E067
-C1E068
-C1E069
-C1E070
-C1E071
-C1E072
-C1E073
-C1E074
-C1E075
-C1E076
-C1E077
-C1E078
-C1E079
-C1E080
-C1E081
-C1E082
-C1E083
-C1E084
-C1E085
-C1E086
-C1E087
-C1E088
-C1E089
-C1E090
-C1E091
-C1E092
-C1E093
-C1E094
-C1E095
-C1E096
-C1E097
-C1E098
-C1E099
-C1E100
-C1E101
-C1E102
-C1E103
-C1E104
-C1E105
-C2E001
-C2E002
-C2E003
-C2E004
-C2E005
-C2E006
-C2E007
-C2E008
-C2E009
-C2E010
-C2E011
-C2E012
-C2E013
-C2E014
-C2E015
-C2E016
-C2E017
-C2E018
-C2E019
-C2E020
-C2E021
-C2E022
-C2E023
-C2E024
-C2E025
-C2E026
-C2E027
-C2E028
-C2E029
-C2E030
-C2E031
-C2E032
-C2E033
-C2E034
-C2E035
-C2E036
