--- data/val_files
@@ -1,4 +1,0 @@
-C1E106
-C1E107
-C2E037
-C2E038
