--- scripts/train.yaml
@@ -1,2 +1,0 @@
-batch_size: 8
-tensorboard_logdir: '../tensorboard_logdir/mistral-test-1/'
