--- src/CRD3_summarization/datasets/__init__.py
@@ -1,1 +1,0 @@
-from .CRD3Datasets import MistralliteCRD3Dataset, MistralCRD3Dataset, EncoderDecoderCRD3Dataset
