--- src/CRD3_summarization/models/__init__.py
@@ -1,3 +1,0 @@
-# TODO: Fix
-# from CRD3SummarizationModel import CRD3SummarizationModel
-# from Trainer import Trainer
