--- tests/test_dataset.py
@@ -1,9 +1,0 @@
-import pytest
-
-
-def test_crd3_encoder_decoder_dataset():
-    ...
-
-
-def test_crd3_decoder_only_dataset():
-    ...
